#21 Add DeliveryModifier

+++ truck/src/main/java/com/polytech/si5/al/dronedelivery/team/g/truck/entities/Position.java
@@ -5,29 +5,29 @@ import javax.persistence.Embeddable;
 @Embeddable
 public class Position {
 
-    private int latitude;
-    private int longitude;
+    private double latitude;
+    private double longitude;
 
     public Position(){}
 
-    public Position(int latitude, int longitude) {
+    public Position(double latitude, double longitude) {
         this.latitude = latitude;
         this.longitude = longitude;
     }
 
-    public int getLatitude() {
+    public double getLatitude() {
         return latitude;
     }
 
-    public void setLatitude(int latitude) {
+    public void setLatitude(double latitude) {
         this.latitude = latitude;
     }
 
-    public int getLongitude() {
+    public double getLongitude() {
         return longitude;
     }
 
-    public void setLongitude(int longitude) {
+    public void setLongitude(double longitude) {
         this.longitude = longitude;
     }
